# server/api/openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: CipherScan API v1
  version: 1.0.0-preview
  summary: >-
    Versioned, contract-stable read/write API for the Zcash blockchain data and privacy-intelligence
    features exposed by CipherScan.
  description: >-
    This is the v1 contract layer. It is gated behind API_V1_ENABLED and, pre-launch, an
    X-API-Preview-Key header (see the previewKey security scheme) — see server/api/v1/README.md in
    the repository for exact mount instructions and current caveats. Every success response uses the
    {data, meta} envelope; every error response is an RFC 9457 problem+json document. Every public
    inventory entry is adapted; private, internal, operational, and deprecated surfaces are
    excluded.
  contact:
    name: CipherScan / Atmosphere Labs
servers:
  - url: https://api.mainnet.cipherscan.app
    description: Mainnet API host.
  - url: https://api.testnet.cipherscan.app
    description: Testnet API host.
  - url: http://127.0.0.1:3001
    description: Local development API host.
...
security:
  - previewKey: []
components:
  securitySchemes:
    previewKey:
      type: apiKey
      in: header
      name: X-API-Preview-Key
      description: Required while API_V1_LAUNCHED=false. Dropped once v1 is generally available.
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem+json error body.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI identifying the problem category.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
    Meta:
      type: object
      required:
        - requestId
        - network
        - generatedAt
        - cache
      properties:
        requestId:
          type: string
          format: uuid
        network:
          type: string
          enum:
            - mainnet
            - testnet
            - crosslink-testnet
        generatedAt:
          type: string
          format: date-time
        indexedHeight:
          type:
            - integer
            - 'null'
        cache:
          type: object
          properties:
            status:
              type: string
              enum:
                - hit
                - miss
                - stale
                - unknown
            ageSeconds:
              type:
                - number
                - 'null'
        dataAgeSeconds:
          type: number
        warnings:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
        page:
          type: object
          description: Present on cursor-paginated list endpoints.
          properties:
            limit:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
            nextCursor:
              type:
                - string
                - 'null'
            prevCursor:
              type:
                - string
                - 'null'
            total:
              type:
                - integer
                - 'null'
...
paths:
...
  /v1/transparent/exposed:
    get:
      operationId: get_transparent_Exposed
      summary: Transparent addresses with exposed-linkage risk.
      tags:
        - transparent
      x-cipherscan-legacy-path: /api/transparent/exposed
      x-cipherscan-legacy-method: GET
      x-cipherscan-classification: public
      x-cipherscan-v1-status: adapter
      responses:
        '200':
          description: Transparent addresses with exposed-linkage risk.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    oneOf:
                      - type: object
                      - type: array
                        items: {}
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Validation error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited (enforced by the legacy endpoint this proxies to).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '502':
          description: Upstream (legacy API) error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '504':
          description: Upstream (legacy API) timeout.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/transparent/exposed/summary:
    get:
      operationId: get_transparent_Exposed_Summary
      summary: Summary of transparent exposure risk.
      tags:
        - transparent
      x-cipherscan-legacy-path: /api/transparent/exposed/summary
      x-cipherscan-legacy-method: GET
      x-cipherscan-classification: public
      x-cipherscan-v1-status: adapter
      responses:
        '200':
          description: Summary of transparent exposure risk.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    oneOf:
                      - type: object
                      - type: array
                        items: {}
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Validation error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited (enforced by the legacy endpoint this proxies to).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '502':
          description: Upstream (legacy API) error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '504':
          description: Upstream (legacy API) timeout.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...